chore: Move didconnection under pkg/store/
closes #1047
