+++ pkg/store/did/didconnection.go
@@ -4,7 +4,7 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package didconnection
+package did
 
 import (
 	"encoding/json"
